Feat scene serialization (#150)
* Changed editor info and editor camera component fetches to use references. Fixed mouseInScene bug

* Started mesh rendering refactor

* Formatted

* Fixed NDC mapping

* Refactored more of the material system, aided by AI

* Refactored grid and mesh material system to ensure proper ordering, this commit was made with the assistance of AI

* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Feat buffer resizing (#143)

* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Meshes working with textures

* Use normals by default on GLTF load and correct directional light usage

* Formatted files

* Fixed app not closing on quit event

* Fixed text input not working

* Fixed wheel acceleration

* Fixed command panel search entity command by refactoring it to a general function

* Fixed find entity and add component popups

* Feat mesh textures (#145)

* Created dynamic buffers by resizing

* Fixed memory read bounds and formatting

* Fixed comments on PR

* Formatted

* Meshes working with textures

* Use normals by default on GLTF load and correct directional light usage

* Formatted files

* Fixed ImGizmo

* Fixed some comments

* Moved editor camera to comp ref

* Fixed most of the comments

* Formatted

* Merged with dev and fixed issues

* Updated CMakeLists

* Fixed transformation storage

* Started on material editor fixes

* Material property flags on editor implemented

* Fixed clicking on hierarchy panel and translation snapping

* Font changed to IBMPlexSeriff, AI aided

* Forgot to commit header

* First steps towards real serialization

* Before changes to vector visitor

* First working serialization

* Adding serializers

* Basic serialization of components finished

* Worked ignoring optional properties on JSON

* Fixed transform serialization, ignoring internal errors of JSON deserializer

* In the middle of refactoring mesh loading for better serialization

* Meta files for one model

* Before changing Ref to use 32bit unsigned integers

* Changed Ref's internal hash to be 32-bit

* Working single mesh exporter

* Working single mesh with textures

* Updated cmake files

+++ src/engine_core/assets/src/HMeta.hpp
@@ -17,7 +17,7 @@ namespace Hush
 
 		uint16_t version = VERSION;
 		uint32_t id = 0;
-		uint64_t sourceHash = 0;
+		uint32_t sourceHash = 0;
 		uint64_t sourceMTime = 0;
 		std::string assetType; // "texture" | "shader" | "model" | "unknown"
